ai-adk-mcp-cloud-run-flightbooking/mcp_agent/requirements.txt
google-adk==1.1.1
google-cloud-bigquery==3.34.0
google-cloud-aiplatform[adk,agent_engines]==1.98.0
mcp[cli]==1.9.4
fastmcp==2.9.2